Update ATS_EX.ino
More refined pass with FM frequency tuning, seek up, seek down, volume control and mute are all implemented and appear to function correctly.

Not implemented is AM tuning, SB tuning and changing frequency bands if not on FM (or one of the other bands). Not sure that is necessary...but with a band+- button I think it is. Not my focus for the moment though...