Add files via upload
Unfortunately you cannot connect the USB UART ports to each other because they are both slaves. And unfortunately ESPHome doesn't support OTG USB_CDC in order to create a host. This can be done in ESP-IDF but I am not going to try. Bypassing the USB ports and just creating a UART direct wiring was my final solution. One other solution is to remove the UART to USB ICs on both boards and carefully wire the UART lines directly to the USB lines (making sure to cross your TX/RX lines) making the USB into a UART. Of course if you do that, make sure NOT to use USB communications with either of those ports anymore.