Update ATS_EX.ino
Final pass for FM only functionality. Modifications will need to be done if you want to do AM and SBW control.

Added feedback if/when volume or station is changed locally. Or...really, feedback for when a scan up/down is taking place. Now if you trigger a scan up/down the stations will be sent back to the ESP32 which then sends it back to home assistant keeping both in sync.

Unfortunately feedback required the battery level to be removed to make space for the additional code. So if you add the battery level feature back, it will remove the feedback code automatically to make sure it still builds (so keep that in mind).