Update defs.h
Had to add conditional build defines that removed the battery voltage display if/when serial control is being used. The feedback part of the serial control is too large to fit both (without possibly rewriting the battery voltage part which looks excessive). Anyway, if you decide to show the battery voltage regardless, the conditional build will automatically remove the feedback feature which keeps the radio in sync with home assistant when local changes or a scan takes place.